moved the api routes and html routes to individual files
